Package renamed to bongpy instead of django_dynaconf in order to avoid confusion with dynaconf
