+++ README.md
@@ -1,10 +1,10 @@
-# django-dynaconf
+# django-bongpy
 
-[![Build](https://github.com/mhsiddiqui/django-dynaconf/actions/workflows/test.yml/badge.svg?branch=main)](https://github.com/mhsiddiqui/django-dynaconf/actions/workflows/test.yml) [![codecov](https://codecov.io/gh/mhsiddiqui/django-dynaconf/branch/main/graph/badge.svg?token=FFXY6AZEDQ)](https://codecov.io/gh/mhsiddiqui/django-dynaconf)
+[![Build](https://github.com/mhsiddiqui/django-bongpy/actions/workflows/test.yml/badge.svg?branch=main)](https://github.com/mhsiddiqui/django-bongpy/actions/workflows/test.yml) [![codecov](https://codecov.io/gh/mhsiddiqui/django-bongpy/branch/main/graph/badge.svg?token=FFXY6AZEDQ)](https://codecov.io/gh/mhsiddiqui/django-bongpy)
 
 
 
-Django-dynaconf provides dynamic configuration for your Django project which you can control from Django admin. You can change your configurations at any time without any deployment.
+Django-bongpy provides dynamic configuration for your Django project which you can control from Django admin. You can change your configurations at any time without any deployment.
 
 **Note**: You cannot add settings required for Django to operate in this like *INSTALLED_APPS*
 
@@ -16,16 +16,16 @@ Django-dynaconf provides dynamic configuration for your Django project which you
 
 1. Run below command to install.
 
-> pip install django-dynaconf
+> pip install django-bongpy
 
-2. Add `django_dynaconf` in your INSTALLED_APPS.
+2. Add `bongpy` in your INSTALLED_APPS.
 3. Run migration by running following command
 
 > python manage.py migrate
 
 # Usage
 _____
-Just go to your admin dashboard, in the Dynaconf section, add configurations. 
+Just go to your admin dashboard, in the Bongpy section, add configurations. 
 
 ## Listing
 ![image info](./images/listing.png)
@@ -88,7 +88,7 @@ Time should be in `HH:MM:SS` format and it will be converted to `datetime.time` 
 In your code, you can use it like this.
 
 ```python
-from django_dynaconf.configs import configs
+from bongpy.configs import configs
 
 print(configs.KEY_OF_YOUR_CONFIGURATION)
 ```
@@ -100,17 +100,17 @@ At any time, when you will change your config value or add a new value, it will 
 You can group configurations. Configurations of a particular group can be accessed like this.
 
 ```python
-from django_dynaconf.configs import configs
+from bongpy.configs import configs
 
 print(configs.grouped(group='NAME_OF_GROUP'))
 ```
 
 ## Default value
 
-It is quite possible when you are using a configuration in your code and you have not added that configuration yet. In this case it is best to define a default value of that configuration in your settings so that that default value will be used instead of an error. To define default value of your configurations, define `DYNACONF_DEFAULTS` in settings like this
+It is quite possible when you are using a configuration in your code and you have not added that configuration yet. In this case it is best to define a default value of that configuration in your settings so that that default value will be used instead of an error. To define default value of your configurations, define `BONGPY_DEFAULTS` in settings like this
 
 ```python
-DYNACONF_DEFAULTS = {
+BONGPY_DEFAULTS = {
     "KEY_1": True,
     "KEY_2": 2000
 }
