+++ bongpy/apps.py
@@ -4,5 +4,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class DjangoDynaconfConfig(AppConfig):
-    name = 'django_dynaconf'
+class BongPyConfig(AppConfig):
+    name = 'bongpy'
