+++ tox.ini
@@ -39,7 +39,7 @@ commands =
     coverage xml
 
 [testenv:flake8-py39]
-commands = flake8 django_dynaconf
+commands = flake8 bongpy
 deps = flake8
 
 [flake8]
